+++ custom_components/worx_vision_cloud/lawn_mower.py
@@ -14,16 +14,20 @@ from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from .const import DOMAIN
 from .entity import WorxVisionEntity
-from .helpers import get_dict_value, rtk_at_station, rtk_distance_to_station_m
+from .helpers import (
+    DOCKED_STATUS_IDS,
+    ERROR_STATUS_IDS,
+    MOWING_STATUS_IDS,
+    PAUSED_STATUS_IDS,
+    RETURNING_STATUS_IDS,
+    STARTING_STATUS_IDS,
+    get_dict_value,
+    rtk_at_station,
+    rtk_distance_to_station_m,
+)
 
 _LOGGER = logging.getLogger(__name__)
 
-MOWING_STATUS_IDS = {7, 8, 12, 32, 110, 111}
-RETURNING_STATUS_IDS = {4, 5, 6, 30, 104}
-STARTING_STATUS_IDS = {2, 3, 33, 103}
-PAUSED_STATUS_IDS = {34}
-DOCKED_STATUS_IDS = {1}
-ERROR_STATUS_IDS = {9, 10, 13}
 RAIN_DELAY_ERROR_DESCRIPTIONS = {"rain delay", "rain_delay"}